move generic list to basic components dir
